* CONNECTED FooCommerce.SqlScripts with the Test DB + ADDED More tests for services `LocationService` and `UserService` + ADDED a concept of entities as C# classes to be usable through the codes

+++ FooCommerce.Application/Enums/Membership/LocationDivisions.cs
@@ -1,6 +1,6 @@
 ﻿namespace FooCommerce.Application.Enums.Membership;
 
-public enum LocationDivisions : ushort
+public enum LocationDivisions : byte
 {
     Country = 0,
     State = 1,
