+++ FooCommerce.Application/Enums/Membership/UserCommunicationTypes.cs
@@ -1,6 +1,6 @@
 ﻿namespace FooCommerce.Application.Enums.Membership;
 
-public enum UserCommunicationTypes : ushort
+public enum UserCommunicationTypes : byte
 {
     Email = 0,
     Mobile = 1,
