+++ FooCommerce.Application/Enums/Membership/UserInformationTypes.cs
@@ -1,6 +1,6 @@
 ﻿namespace FooCommerce.Application.Enums.Membership;
 
-public enum UserInformationTypes : ushort
+public enum UserInformationTypes : byte
 {
     Name = 0,
     Surname = 1
